[OSDEV-2867] Move OriginSourceMiddleware into api/middlewares
Extract OriginSourceMiddleware and is_health_check_request into
api/middlewares/ (origin_source.py + utils.py) so shared probe
skipping stays reusable without bloating api/middleware.py.
